fix: move test to good directory
